# Speed up TestIntegration in campaigns package