--- a/enterprise/internal/campaigns/store_test.go
+++ b/enterprise/internal/campaigns/store_test.go
@@ -35,7 +35,7 @@ func testStore(db *sql.DB) func(*testing.T) {
 		// Create a test repo
 		reposStore := repos.NewDBStore(db, sql.TxOptions{})
 		repo := &repos.Repo{
-			Name: "github.com/sourcegraph/sourcegraph",
+			Name: "github.com/sourcegraph/sourcegraph-test-repo",
 			ExternalRepo: api.ExternalRepoSpec{
 				ID:          "external-id",
 				ServiceType: "github",
@@ -2795,7 +2795,7 @@ func testStore(db *sql.DB) func(*testing.T) {
 	}
 }
 
-func testProcessChangesetJob(db *sql.DB) func(*testing.T) {
+func testProcessChangesetJob(db *sql.DB, userID int32) func(*testing.T) {
 	return func(t *testing.T) {
 		now := time.Now().UTC().Truncate(time.Microsecond)
 		clock := func() time.Time { return now.UTC().Truncate(time.Microsecond) }
@@ -2804,7 +2804,7 @@ func testProcessChangesetJob(db *sql.DB) func(*testing.T) {
 		// Create a test repo
 		reposStore := repos.NewDBStore(db, sql.TxOptions{})
 		repo := &repos.Repo{
-			Name: "github.com/sourcegraph/sourcegraph",
+			Name: "github.com/sourcegraph/changeset-job-test",
 			ExternalRepo: api.ExternalRepoSpec{
 				ID:          "external-id",
 				ServiceType: "github",
@@ -2821,15 +2821,9 @@ func testProcessChangesetJob(db *sql.DB) func(*testing.T) {
 			t.Fatal(err)
 		}
 
-		var userID int32
-		err := db.QueryRow("INSERT INTO users (username) VALUES ('admin') RETURNING id").Scan(&userID)
-		if err != nil {
-			t.Fatal(err)
-		}
-
 		s := NewStoreWithClock(db, clock)
 		patchSet := &cmpgn.PatchSet{UserID: userID}
-		err = s.CreatePatchSet(context.Background(), patchSet)
+		err := s.CreatePatchSet(context.Background(), patchSet)
 		if err != nil {
 			t.Fatal(err)
 		}