--- a/enterprise/internal/campaigns/webhooks_test.go
+++ b/enterprise/internal/campaigns/webhooks_test.go
@@ -38,7 +38,7 @@ import (
 var update = flag.Bool("update", false, "update testdata")
 
 // Run from integration_test.go
-func testGitHubWebhook(db *sql.DB) func(*testing.T) {
+func testGitHubWebhook(db *sql.DB, userID int32) func(*testing.T) {
 	return func(t *testing.T) {
 		now := time.Now().UTC().Truncate(time.Microsecond)
 		clock := func() time.Time { return now }
@@ -47,14 +47,10 @@ func testGitHubWebhook(db *sql.DB) func(*testing.T) {
 
 		rcache.SetupForTest(t)
 
+		truncateTables(t, db, "changeset_jobs", "changeset_events", "changesets")
+
 		cf, save := newGithubClientFactory(t, "github-webhooks")
 		defer save()
-
-		var userID int32
-		err := db.QueryRow("INSERT INTO users (username) VALUES ('admin') RETURNING id").Scan(&userID)
-		if err != nil {
-			t.Fatal(err)
-		}
 
 		secret := "secret"
 		repoStore := repos.NewDBStore(db, sql.TxOptions{})
@@ -69,7 +65,7 @@ func testGitHubWebhook(db *sql.DB) func(*testing.T) {
 			}),
 		}
 
-		err = repoStore.UpsertExternalServices(ctx, extSvc)
+		err := repoStore.UpsertExternalServices(ctx, extSvc)
 		if err != nil {
 			t.Fatal(t)
 		}
@@ -133,14 +129,7 @@ func testGitHubWebhook(db *sql.DB) func(*testing.T) {
 			_, name := path.Split(fixtureFile)
 			name = strings.TrimSuffix(name, ".json")
 			t.Run(name, func(t *testing.T) {
-				_, err = db.Exec("ALTER SEQUENCE changeset_events_id_seq RESTART")
-				if err != nil {
-					t.Fatal(err)
-				}
-				_, err = db.Exec("TRUNCATE TABLE changeset_events")
-				if err != nil {
-					t.Fatal(err)
-				}
+				truncateTables(t, db, "changeset_events")
 
 				tc := loadWebhookTestCase(t, fixtureFile)
 
@@ -201,7 +190,7 @@ func testGitHubWebhook(db *sql.DB) func(*testing.T) {
 }
 
 // Run from integration_test.go
-func testBitbucketWebhook(db *sql.DB) func(*testing.T) {
+func testBitbucketWebhook(db *sql.DB, userID int32) func(*testing.T) {
 	return func(t *testing.T) {
 		now := time.Now().UTC().Truncate(time.Microsecond)
 		clock := func() time.Time { return now }
@@ -210,14 +199,10 @@ func testBitbucketWebhook(db *sql.DB) func(*testing.T) {
 
 		rcache.SetupForTest(t)
 
+		truncateTables(t, db, "changeset_jobs", "changeset_events", "changesets")
+
 		cf, save := newGithubClientFactory(t, "bitbucket-webhooks")
 		defer save()
-
-		var userID int32
-		err := db.QueryRow("INSERT INTO users (username) VALUES ('admin') RETURNING id").Scan(&userID)
-		if err != nil {
-			t.Fatal(err)
-		}
 
 		secret := "secret"
 		repoStore := repos.NewDBStore(db, sql.TxOptions{})
@@ -234,7 +219,7 @@ func testBitbucketWebhook(db *sql.DB) func(*testing.T) {
 			}),
 		}
 
-		err = repoStore.UpsertExternalServices(ctx, extSvc)
+		err := repoStore.UpsertExternalServices(ctx, extSvc)
 		if err != nil {
 			t.Fatal(t)
 		}
@@ -308,14 +293,7 @@ func testBitbucketWebhook(db *sql.DB) func(*testing.T) {
 			_, name := path.Split(fixtureFile)
 			name = strings.TrimSuffix(name, ".json")
 			t.Run(name, func(t *testing.T) {
-				_, err = db.Exec("ALTER SEQUENCE changeset_events_id_seq RESTART")
-				if err != nil {
-					t.Fatal(err)
-				}
-				_, err = db.Exec("TRUNCATE TABLE changeset_events")
-				if err != nil {
-					t.Fatal(err)
-				}
+				truncateTables(t, db, "changeset_events")
 
 				tc := loadWebhookTestCase(t, fixtureFile)