Before:

    $ go test -count=1 .
    ok      github.com/sourcegraph/sourcegraph/enterprise/internal/campaigns
    7.521s

    $ go test -count=1 .
    ok      github.com/sourcegraph/sourcegraph/enterprise/internal/campaigns

    7.239s
    $ go test -count=1 .
    ok      github.com/sourcegraph/sourcegraph/enterprise/internal/campaigns
    7.269s

    $ go test -count=1 .
    ok      github.com/sourcegraph/sourcegraph/enterprise/internal/campaigns
    7.487s

After:

    $ go test -count=1 .
    ok      github.com/sourcegraph/sourcegraph/enterprise/internal/campaigns
    4.266s

    $ go test -count=1 .
    ok      github.com/sourcegraph/sourcegraph/enterprise/internal/campaigns
    4.367s

    $ go test -count=1 .
    ok      github.com/sourcegraph/sourcegraph/enterprise/internal/campaigns
    4.298s



<!-- Reminder: Have you updated the changelog and relevant docs (user docs, architecture diagram, etc) ? -->